JWTHelperService no longer dependent on external configuration
This patch sets a default value to the JWTHelperServices' constructor, while
still requiring that it be provided with an injectee. This achieves two things:
First, it allows the JWTHelperService to be used with no constructor arguments,
should someone only want to use the utility nature of the class. Secondly,
it preserves the injection requirement, so that someone who chooses to use the
injection use case is still notified of the required external contract.

Documentation has also been updated.